Refs #126. Moved test case to a better location
