+++ tests/test_wordml.py
@@ -6,34 +6,8 @@ from __future__ import (
 
 import unittest
 
-from pydocx.packaging import ZipPackage
 from pydocx.util.xml import xml_tag_split
 from pydocx.openxml.packaging import MainDocumentPart, WordprocessingDocument
-
-
-class ZipPackageTestCase(unittest.TestCase):
-    def setUp(self):
-        self.package = ZipPackage(
-            path='tests/fixtures/no_break_hyphen.docx',
-        )
-
-    def test_relationship_uri(self):
-        self.assertEqual(
-            self.package.relationship_uri,
-            '/_rels/.rels',
-        )
-
-    def test_relationship_part_exists(self):
-        assert self.package.part_exists(self.package.relationship_uri)
-
-    def test_word_document_part_exists(self):
-        assert self.package.part_exists('/word/document.xml')
-
-    def test_package_relationship_part_stream(self):
-        part = self.package.get_part('/_rels/.rels')
-        data = part.stream.read()
-        assert data
-        assert data.startswith(b'<?xml version="1.0" encoding="UTF-8"?>')
 
 
 class WordprocessingDocumentTestCase(unittest.TestCase):
